backend/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
razorpay==1.3.0
Pillow==9.4.0
python-dotenv==1.0.0
pytest==7.3.1
